Sigo con el SET

+++ libraries/generales/generales.h
@@ -12,7 +12,7 @@
 #include <stdlib.h>
 
 int min(int, int);
-int roundNumber(double);
+int roundNumber(float);
 t_link_element* list_find_element_with_param(t_list*, void*, int(*condition)(void*, void*), int*);
 void* list_find_with_param(t_list*, void*, int(*condition)(void*, void*));
